chore: OtherPeopleViewController에 BrowseStyleDataManager로 데이터 받아오는 방식으로 수정. OtherPeopleView 삭제
