+++ HowsTheWear/Sources/OtherPeopleScene/OtherPeopleMainScene/Controllers/OtherPeopleViewController.swift
@@ -9,16 +9,39 @@ import UIKit
 
 final class OtherPeopleViewController: UIViewController {
     
-    private let otherPeopleImageArray: [[UIImage]] = [[]]
+    private let browseStyleDataManager = BrowseStyleDataManager()
     
-    private let otherPeopleView = OtherPeopleView()
+    private var dataArray:[[BrowseStyleDataModel]] = []
+    
+//    private let otherPeopleView = OtherPeopleView()
+    
+    private var hashTagNumber: Int = 5
+    
+    private var hashTagTitleArray: [String] = ["#캐쥬얼", "#스트릿", "#페미닌", "#미니멀"]
+    
+    let othersPeopleCollectionView = BrowseMainCollectionView(isHiddenCellUserID: false, cellImageCorenerRadius: 8)
+    
+    let myPostButton = UIButton().then {
+        $0.backgroundColor = .black
+        $0.setImage(UIImage(systemName: "plus")?
+            .withConfiguration(
+                UIImage.SymbolConfiguration(pointSize: 20, weight: .regular)), for: .normal)
+        $0.tintColor = .white
+    }
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        view = otherPeopleView
+//        view = otherPeopleView
         configureInitialSetting()
         configureCollectionView()
         configureAction()
+        configureSubview()
+        configureLayout()
+    }
+    
+    override func viewDidLayoutSubviews() {
+        super.viewDidLayoutSubviews()
+        configurePlusButton()
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -29,7 +52,7 @@ final class OtherPeopleViewController: UIViewController {
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         tabBarController?.tabBar.isHidden = false
-        otherPeopleView.myPostButton.isHidden = false
+        myPostButton.isHidden = false
     }
     
     override func viewWillDisappear(_ animated: Bool) {
@@ -37,7 +60,7 @@ final class OtherPeopleViewController: UIViewController {
         if let navigationController = navigationController,
            navigationController.topViewController is MyPostViewController
         {
-            otherPeopleView.myPostButton.isHidden = true
+            myPostButton.isHidden = true
             tabBarController?.tabBar.isHidden = true
         }
     }
@@ -51,6 +74,13 @@ extension OtherPeopleViewController {
     private func configureInitialSetting() {
         view.backgroundColor = #colorLiteral(red: 0.9803921569, green: 0.9803921569, blue: 0.9803921569, alpha: 1)
         configureNaviBar()
+        dataArray = [
+            browseStyleDataManager.fetchThisWeekImagesData(),
+            browseStyleDataManager.fetchNextWeekImagesData(),
+            browseStyleDataManager.fetchLastYearImagesData(),
+            browseStyleDataManager.fetchThisWeekImagesData(),
+            browseStyleDataManager.fetchNextWeekImagesData(),
+        ]
     }
     
     private func configureNaviBar() {
@@ -83,24 +113,41 @@ extension OtherPeopleViewController {
         navigationItem.backBarButtonItem = backBarButton
     }
     
+    private func configurePlusButton() {
+        myPostButton.layer.cornerRadius = myPostButton.frame.height / 2
+        myPostButton.layer.shadowColor = UIColor.black.cgColor
+        myPostButton.layer.shadowOpacity = 0.5
+        myPostButton.layer.shadowOffset = CGSize(width: 0, height: 0)
+        myPostButton.layer.shadowRadius = 10
+    }
+    
 }
 
 // MARK: - Configure CollectionView
 
 extension OtherPeopleViewController {
-    
+
+    func setCollectionViewCellSelectionHandler(_ handler: @escaping ((IndexPath) -> Void)) {
+        othersPeopleCollectionView.didSelectCell = handler
+    }
     
     private func configureCollectionView() {
-        otherPeopleView.othersPeopleCollectionView.delegate = self
+        othersPeopleCollectionView.delegate = self
         
-        otherPeopleView.setCollectionViewCellSelectionHandler { [weak self] indexPath in
+        setCollectionViewCellSelectionHandler { [weak self] indexPath in
             let detailViewController = OtherPeopleDetailViewController()
             self?.navigationController?.pushViewController(detailViewController, animated: true)
             
 //            // 모델, 데이터매니저 구현 후 데이터 받아오는 메서드 작성예정
         }
         
+        othersPeopleCollectionView.configureContents(
+            sectionCount: hashTagNumber,
+            data: dataArray,
+            sectionTitles: hashTagTitleArray
+        )
     }
+    
 }
 
 // MARK: - Implement CollectionView Delegate
@@ -120,7 +167,7 @@ extension OtherPeopleViewController: UICollectionViewDelegate, browseCollectionR
 
 extension OtherPeopleViewController {
     private func configureAction() {
-        otherPeopleView.myPostButton.addTarget(self, action: #selector(myPostButtonDidTap), for: .touchUpInside)
+        myPostButton.addTarget(self, action: #selector(myPostButtonDidTap), for: .touchUpInside)
     }
     
     @objc private func myPostButtonDidTap() {
@@ -128,6 +175,34 @@ extension OtherPeopleViewController {
         navigationController?.pushViewController(myPostVC, animated: true)
         navigationController?.navigationBar.tintColor = .black
         navigationController?.navigationBar.topItem?.title = ""
+    }
+    
+}
+
+
+// MARK: - Configure UI
+
+extension OtherPeopleViewController {
+    private func configureSubview() {
+        [othersPeopleCollectionView, myPostButton].forEach {
+            view.addSubview($0)
+        }
+    }
+    
+    private func configureLayout() {
+        let safeArea = view.safeAreaLayoutGuide
+        othersPeopleCollectionView.snp.makeConstraints { make in
+            make.top.equalTo(safeArea.snp.top).inset(30)
+            make.bottom.equalTo(safeArea.snp.bottom)
+            make.leading.equalTo(safeArea.snp.leading).inset(20)
+            make.trailing.equalTo(safeArea.snp.trailing)
+        }
+        
+        myPostButton.snp.makeConstraints {
+            $0.right.equalTo(-15)
+            $0.bottom.equalTo(safeArea.snp.bottom).inset(15)
+            $0.width.height.equalTo(55)
+        }
     }
     
 }
